- update recommend status to generating after creating a recommend task. - move the type definition to /models to avoid "ReferenceError: Cannot access 'IbisAdaptor' before initialization "
